+++ q2gui/pyqt6/widgets/q2date.py
@@ -35,7 +35,7 @@ class q2date(QComboBox, Q2Widget):
         self.lineedit.setStyleSheet("QLineEdit")
         self.set_text(self.meta.get("data"))
         # self.setMinimumWidth(QFontMetrics(self.font()).width("0") * 14)
-        self.setMinimumWidth(QFontMetrics(self.font()).horizontalAdvance("0") * 14)
+        self.setMinimumWidth(int(QFontMetrics(self.font()).horizontalAdvance("0") * 14))
         self.setSizePolicy(QSizePolicy.Policy.Maximum, QSizePolicy.Policy.Maximum)
 
         class q2DateValidator(QValidator):
