+++ q2gui/pyqt6/widgets/q2label.py
@@ -33,7 +33,7 @@ class q2label(QLabel, Q2Widget):
         # super().__init__(meta)
         self.set_text(self.meta["label"])
         self.setWordWrap(True)
-        self.set_maximum_height(self.get_default_height() * 1.5)
+        self.set_maximum_height(int(self.get_default_height() * 1.5))
 
     def set_style_sheet(self, style_text):
         super().set_style_sheet(style_text)
